test(auth): fold enforce-scope parity cases into shadowAuth servertest
